OF: overlay: Skip overlay_removal_is_ok check for PL use cases
The existing overlay framework allows the removal of overlay's only
if it's the top-most one. This limitation puts restrictions on PL
dynamic configuration use cases. If the PL design has 2 RP regions,
PR1 is loaded, followed by PR2 the existing overlay framework does not
allow the removal of PR1 first. It force us to follow the LIFO order
to remove the overlays. As a workaround solution to avoid this limitation,
Bypass the overlay_removal_is_ok check if CONFIG_OF_FPGA_REGION enabled.
It will allow us to load and unload PL images independently from the
targeted RP region.

Signed-off-by: Nava kishore Manne <nava.kishore.manne@amd.com>
Acked-by : Siva Durga Prasad Paladugu <siva.durga.prasad.paladugu@amd.com>
Signed-off-by: Radhey Shyam Pandey <radhey.shyam.pandey@amd.com>