xilinx-ai-engine: Add new api to clear the core regs
Add api to clear the core regs. core_regs_clr has list of registers that
needs to be zeroized. Some of the consecutive registers are not
consecutive and have holes in them. step and width is used to detect
this. When width == step, this means the register list is linear we can
use memset_io for this memory range. If it is not linear, write*_relaxed
is used with step increment in address.

Signed-off-by: Govindarajulu Varadarajan <govindarajulu.varadarajan@amd.com>
Acked-by: Gregory Williams <gregory.williams@amd.com>
Message-ID: <20240927055451.72666-4-govindarajulu.varadarajan@amd.com>